requirements.txt
fastapi==0.115.0
uvicorn==0.30.6
pydantic==2.9.2
python-dotenv==1.0.1
httpx==0.27.2
orjson==3.10.7
tqdm==4.66.5
beautifulsoup4==4.12.3
markdownify==0.13.1
GitPython==3.1.43
faiss-cpu
nltk==3.9.1
regex==2024.7.24
datasets==2.21.0
transformers==4.44.2
accelerate==0.34.2
peft==0.12.0
einops==0.8.0
sentence-transformers==3.1.1
